+++ changelog.md
@@ -1,6 +1,7 @@
-#### 0.0.5 (2021-09-27)
+#### 0.0.5 (2021-09-28)
 
-- add clearer redirect logic (#1b2cec7)
+- fix for `remove_default_language_from_url` (#5d24b90)
+- add clearer redirect logic (#a75f634)
 - guess requested post types from `wp_query` (#218858b)
 - guess post_type in taxonomy archives (#d734a61)
 - optimize `convert_url`  for edge  cases where the a WP_Post can have various URLs (#72b7496)